put all methods regulating the game into Game (removing PlayerActions)
